Move find module up one level.
