+++ model_metadata/find.py
@@ -3,7 +3,7 @@ import os
 
 from scripting.contexts import cd
 
-from ..errors import MetadataNotFoundError
+from .errors import MetadataNotFoundError
 
 _METADATA_FILES = {
     "api.yaml",
