--- model_metadata/metadata/__init__.py
@@ -1,5 +1,0 @@
-
-from .find import find_model_data_files
-
-
-__all__ = ["find_model_data_files"]
