+++ model_metadata/modelmetadata.py
@@ -7,7 +7,7 @@ import yaml
 
 from .errors import MissingSectionError, MissingValueError
 from .load import load_yaml_file
-from .metadata.find import find_metadata_files
+from .find import find_metadata_files
 from .model_info import ModelInfo
 from .model_parameter import parameter_from_dict, setup_yaml_with_canonical_dict
